requirements.txt
google-generativeai==0.8.4
pymongo==4.11
requests==2.32.3
python-telegram-bot==21.10
python-dotenv==1.0.1